feat: move components into directories; all component styling using @emotion
