+++ src/components/Board.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './index.css';
+import '../styles/index.css';
 import styled from '@emotion/styled/macro';
 import Square from './Square';
 import calculateWinner from './calculateWinner';
