+++ src/components/calculateWinner.js
