+++ src/styles/index.css
@@ -1,0 +1,32 @@
+body {
+  font: 14px "Century Gothic", Futura, sans-serif;
+  margin: 20px;
+}
+
+ol, ul {
+  padding-left: 30px;
+}
+
+
+
+.status {
+  margin-bottom: 10px;
+}
+
+.kbd-navigation .square:focus {
+  background: #ddd;
+}
+
+.game {
+  margin: 0 auto;
+  display: flex;
+  flex-direction: row;
+}
+
+.game-board {
+  margin: 0 auto;
+}
+
+.game-info {
+  margin-left: 20px;
+}
